Add Resolver.
Since we use Expressions as keys to lookup variables we need Expr
objects to be hashable. There was a bug where variables on the same name
with the same line would shadow each other, as they would hash to the
same value. This meant it wasn't possible to resolve for statements as
the VarExpr and other references to the var would be on the same line.

To combat this a column parameter was added to the Token object, which
means that all the Scanner tests needed to be updated.